+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - show user-visible error in fullscreen media view if mimetype isn't set or not supported
 
 ### Fixed
+- rename `--debug` flag to `--devmode` (in order to fix #2315)
 - fix duplicated contacts in search
 
 ## [1.20.3] - 2021-06-30